requirements.txt
dj-database-url==0.4.2
Django==1.10.6
djangorestframework==3.6.1
gunicorn==19.7.0
psycopg2==2.6.2
pytest==3.0.4
pytest-django==3.1.2
whitenoise==3.3.0
